python-php-search-start-1
Begun the process of creating the XML, python, and php query tools. Added some files to convert python objects into json for more efficient searching - assuming that the database does not change it will work, otherwise the generator will need to be run every time there's a change but this can work well.